+++ src/main/kotlin/stellarwitch7/libstellar/registry/codec/CodecRegistrar.kt
@@ -1,5 +1,6 @@
 package stellarwitch7.libstellar.registry.codec
 
+import com.mojang.serialization.Codec
 import com.mojang.serialization.Lifecycle
 import com.mojang.serialization.MapCodec
 import net.minecraft.registry.Registry
@@ -8,8 +9,10 @@ import net.minecraft.registry.SimpleRegistry
 import stellarwitch7.libstellar.registry.Registrar
 
 interface CodecRegistrar<T : CodecTypeProvider<T>> : Registrar<CodecType<T>> {
-    val codec: MapCodec<T>
-        get() = registry.codec.dispatchMap(CodecTypeProvider<T>::type, CodecType<T>::codec)
+    val codec: Codec<T>
+        get() = Codec.lazyInitialized {
+            registry.codec.dispatchMap(CodecTypeProvider<T>::type, CodecType<T>::codec).codec()
+        }
 
     fun makeReg(name: String): Registry<CodecType<T>> = SimpleRegistry(RegistryKey.ofRegistry(id(name)), Lifecycle.stable())